Constractors moved to persistence layer.
