+++ app/src/main/java/com/gmail/alexander/taskchronometer/activities/AddEditActivityFragment.java
@@ -17,7 +17,7 @@ import android.widget.Button;
 import android.widget.EditText;
 
 import com.gmail.alexander.taskchronometer.R;
-import com.gmail.alexander.taskchronometer.contractors.TasksContract;
+import com.gmail.alexander.taskchronometer.persistence_layer.contractors.TasksContract;
 import com.gmail.alexander.taskchronometer.domain_layer.Task;
 import com.gmail.alexander.taskchronometer.listeners.OnSaveListener;
