+++ app/src/main/java/com/gmail/alexander/taskchronometer/activities/DurationsReportActivity.java
@@ -20,7 +20,7 @@ import android.widget.DatePicker;
 
 import com.gmail.alexander.taskchronometer.R;
 import com.gmail.alexander.taskchronometer.adapters.DurationsRecViewAdapter;
-import com.gmail.alexander.taskchronometer.contractors.DurationsContract;
+import com.gmail.alexander.taskchronometer.persistence_layer.contractors.DurationsContract;
 
 import java.security.InvalidParameterException;
 import java.util.Date;
