+++ app/src/main/java/com/gmail/alexander/taskchronometer/adapters/DurationsRecViewAdapter.java
@@ -9,7 +9,7 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.gmail.alexander.taskchronometer.R;
-import com.gmail.alexander.taskchronometer.contractors.DurationsContract;
+import com.gmail.alexander.taskchronometer.persistence_layer.contractors.DurationsContract;
 
 import java.util.Locale;
